changing int to long

+++ fichiers modifiés/FpX.c
@@ -2401,7 +2401,7 @@ gener_FpXQ_local(GEN T, GEN p, GEN L)
 /**              FqXn and Newton iterations                           **/
 /**                                                                   **/
 /***********************************************************************/ 
-GEN Fq_ui(int k,GEN T,GEN p)
+GEN Fq_ui(long k,GEN T,GEN p)
 {
   if(T==NULL)
     return mkintn(1,k);
@@ -2415,20 +2415,20 @@ GEN FqX_mul2(GEN P,GEN T,GEN p)
   {
 	  return FqX_mulu(P,2,T,p);
   }
-  int d=lg(P)-3;
-  int k=0;
+  long d=lg(P)-3;
+  long k=0;
   GEN ret=cgetg(lg(P),t_POL);
   for(k=0;k<=d;k++)
     gel(ret,2+k)=Fp_red(shifti(gel(P,2+k),1),p);
   return ret;
 }
 
-GEN FqX_modXn(GEN P,int n,GEN T,GEN p)
+GEN FqX_modXn(GEN P,long n,GEN T,GEN p)
 {	
   if(T!=NULL)
   {
     GEN ret=RgX_copy(P);
-    int k =n;
+    long k =n;
     while(k<lg(P)-2)
     {	
       gel(ret,2+k)=mkpoln(1,gen_0);
@@ -2438,7 +2438,7 @@ GEN FqX_modXn(GEN P,int n,GEN T,GEN p)
     return FqX_red(ret,T,p);
   }
   GEN ret=RgX_copy(P);
-  int k =n;
+  long k =n;
   while(k<lg(P)-2)
   {	
     gel(ret,2+k)=mkintn(1,0);
@@ -2448,13 +2448,13 @@ GEN FqX_modXn(GEN P,int n,GEN T,GEN p)
   return RgX_to_FpX(ret,p);
 }
 static 
-GEN RgXn_mul_basecase(GEN P,GEN Q,int n)
+GEN RgXn_mul_basecase(GEN P,GEN Q,long n)
 {
-  int d =lg(P)+lg(Q)-3;
-  int min =(d>n+2)*(n+2)+(d<=n+2)*d;
+  long d =lg(P)+lg(Q)-3;
+  long min =(d>n+2)*(n+2)+(d<=n+2)*d;
   GEN ret=cgetg(min,t_POL);
-  int k=0;
-  int j;
+  long k=0;
+  long j;
   d=min-3;
 
 
@@ -2479,13 +2479,13 @@ GEN RgXn_mul_basecase(GEN P,GEN Q,int n)
   return ret;
 }
 static 
-GEN RgXn_sqr_basecase(GEN P,int n)
+GEN RgXn_sqr_basecase(GEN P,long n)
 {
-  int d =2*lg(P)-3;
-  int min =(d>n+2)*(n+2)+(d<=n+2)*d;
+  long d =2*lg(P)-3;
+  long min =(d>n+2)*(n+2)+(d<=n+2)*d;
   GEN ret=cgetg(min,t_POL);
-  int k=0;
-  int j;
+  long k=0;
+  long j;
   d=min-3;
 
 	
@@ -2516,10 +2516,10 @@ GEN RgXn_sqr_basecase(GEN P,int n)
   return ret;
 }
 
-static GEN FqXn_mul_karatsuba(GEN P,GEN Q,int n,GEN T,GEN p)
+static GEN FqXn_mul_karatsuba(GEN P,GEN Q,long n,GEN T,GEN p)
 //doesn't reduce coefficients to be faster, they will be reduced in FqXn_mul
 {	
-  int s=RgX_equal(P,mkpoln(1,gen_0))||RgX_equal(Q,mkpoln(1,gen_0));
+  long s=RgX_equal(P,mkpoln(1,gen_0))||RgX_equal(Q,mkpoln(1,gen_0));
   if(T==NULL)
   {
     if(s){return mkpoln(1,gen_0);}
@@ -2534,7 +2534,7 @@ static GEN FqXn_mul_karatsuba(GEN P,GEN Q,int n,GEN T,GEN p)
     if(n<35)
       return RgXn_mul_basecase(P,Q,n);
   }
-  int B=(7*n)/10;
+  long B=(7*n)/10;
   GEN ret;
   GEN P1=RgX_shift(FqX_modXn(P,n,T,p),-B);
   GEN Q1=RgX_shift(FqX_modXn(Q,n,T,p),-B);
@@ -2549,17 +2549,17 @@ static GEN FqXn_mul_karatsuba(GEN P,GEN Q,int n,GEN T,GEN p)
   gerepileupto(avma,ret);
   return ret;
 }
-static GEN FqXn_sqr_karatsuba(GEN P,int n,GEN T,GEN p)
+static GEN FqXn_sqr_karatsuba(GEN P,long n,GEN T,GEN p)
 //doesn't reduce coefficients to be faster, they will be reduced in FqXn_mul
 {	
-  int s=RgX_equal(P,mkpoln(1,gen_0));
+  long s=RgX_equal(P,mkpoln(1,gen_0));
  
 	if(s){return mkpoln(1,Fq_ui(0,T,p));}
 	if (2*degpol(P)  < n) 
 	  return RgX_sqr(P);
 	if(n<35)
 	  return RgXn_sqr_basecase(P,n);
-  int B=(7*n)/10;
+  long B=(7*n)/10;
   GEN ret;
   GEN P1=RgX_shift(FqX_modXn(P,n,T,p),-B);
   GEN P2=FqX_modXn(P,B,T,p);
@@ -2591,10 +2591,10 @@ GEN FqX_div2(GEN P,GEN T,GEN p)
   {
     return FqX_Fq_mul(P,Fq_inv(mkpoln(1,gen_2),T,p),T,p);
   }
-  int d=lg(P)-3;
+  long d=lg(P)-3;
   if(d==-1){return mkpoln(1,gen_0);}
   GEN ret=cgetg(lg(P),t_POL);
-  int k =0;
+  long k =0;
   GEN tmp;
   for(k=0;k<=d;k++)
   {
@@ -2613,11 +2613,11 @@ GEN FqX_div2(GEN P,GEN T,GEN p)
   return ret;
 }
 
-GEN FqX_mulup(GEN P,GEN Q,int n,GEN T,GEN p)
+GEN FqX_mulup(GEN P,GEN Q,long n,GEN T,GEN p)
 // calculate P*Q/x^n (without the low degree coefficients)
 {
 	
-  int d=lg(P)+lg(Q)-5-n;
+  long d=lg(P)+lg(Q)-5-n;
   if(d<=2||d>=110)
   {
     return RgX_shift(FqX_mul(P,Q,T,p),-n);
@@ -2628,9 +2628,9 @@ GEN FqX_mulup(GEN P,GEN Q,int n,GEN T,GEN p)
   ret=RgX_shift(ret,d);
   return ret;
 }
-GEN FqX_sqrup(GEN P,int n,GEN T,GEN p)
+GEN FqX_sqrup(GEN P,long n,GEN T,GEN p)
 {
-  int d=2*lg(P)-5-n;
+  long d=2*lg(P)-5-n;
   if(d<=2||d>=110)
   {
     return RgX_shift(FqX_sqr(P,T,p),-n);
@@ -2642,7 +2642,7 @@ GEN FqX_sqrup(GEN P,int n,GEN T,GEN p)
   ret=RgX_shift(ret,d);
   return ret;
 }
-static GEN FqX_mulup_modxn(GEN P,GEN Q,int t1,int t2,GEN T,GEN p)
+static GEN FqX_mulup_modxn(GEN P,GEN Q,long t1,long t2,GEN T,GEN p)
 // calculate (P*Q/x^t1) mod x^t2 (without the low degree coefficients) 
 {	
   if(t2>110)
@@ -2652,15 +2652,15 @@ static GEN FqX_mulup_modxn(GEN P,GEN Q,int t1,int t2,GEN T,GEN p)
   else
     return FqX_add(FqX_add(FqX_mulup(FqX_modXn(Q,t1,T,p),FqX_modXn(P,t1,T,p),t1,T,p),FqX_add(FqXn_mul(RgX_shift(P,-t1),FqX_modXn(Q,t1,T,p),t2-t1,T,p),FqXn_mul(RgX_shift(Q,-t1),FqX_modXn(P,t1,T,p),t2-t1,T,p),T,p),T,p),RgX_shift(FqXn_mul(RgX_shift(P,-t1),RgX_shift(Q,-t1),t2-2*t1,T,p),t1),T,p);
 }
-GEN FqX_Newton_iteration_inv(GEN I,GEN P,int t1,int t2,GEN T,GEN p)
+GEN FqX_Newton_iteration_inv(GEN I,GEN P,long t1,long t2,GEN T,GEN p)
 //given I=1/P mod x^t1 returns 1/P mod x^t2
 {	
   return FqX_sub(I,RgX_shift(FqXn_mul(FqX_mulup_modxn(I,P,t1,t2,T,p),I,t2-t1,T,p),t1),T,p);
 }
-static int find_size(int l)
+static long find_size(long l)
 {   
-  int ret=1;
-  int tmp=l;
+  long ret=1;
+  long tmp=l;
   while(tmp>1)
   {
       tmp=tmp>>1;
@@ -2669,8 +2669,8 @@ static int find_size(int l)
   if(l==(1<<(ret-1))){ret--;}
   return ret;
 }
-static void find_list(int *list,int l)
-{   int*tmp=(list)+(find_size(l));
+static void find_list(long *list,long l)
+{   long*tmp=(list)+(find_size(l));
     *tmp=l;
     while(l>1)
     {
@@ -2680,7 +2680,7 @@ static void find_list(int *list,int l)
     }
     return;
 }
-GEN FqXn_sqrt(GEN P,int n, GEN T,GEN p)
+GEN FqXn_sqrt(GEN P,long n, GEN T,GEN p)
 // requires find_list,find_size and FqX_Newton_iteration_inv
 {
   if(Fq_issquare(gel(P,2),T,p)==0)
@@ -2689,10 +2689,10 @@ GEN FqXn_sqrt(GEN P,int n, GEN T,GEN p)
   GEN k,i;
   ret=mkpoln(1,Fq_sqrt(gel(P,2),T,p));
   i=mkpoln(1,Fq_inv(Fq_mulu(gel(ret,2),2,T,p),T,p));
-  int t=find_size(n)+1;
-  int tab[t];
+  long t=find_size(n)+1;
+  long tab[t];
   find_list(tab,n);
-  int j=0;
+  long j=0;
   while(j<t-1)
   {	//we compute the inverse and the sqrt at the same time, ret denotes the sqrt and i the inverse        
     k=FqX_sub(RgX_shift(FqX_modXn(P,tab[j+1],T,p),tab[j]-tab[j+1]),FqX_sqrup(ret,tab[j+1]-tab[j],T,p),T,p);
@@ -2704,13 +2704,13 @@ GEN FqXn_sqrt(GEN P,int n, GEN T,GEN p)
   gerepileupto(avma,ret);
   return ret;  
 }
-GEN FqXn_inv(GEN P, int n,GEN T,GEN p)
+GEN FqXn_inv(GEN P, long n,GEN T,GEN p)
 {  
   GEN ret=mkpoln(1,Fq_inv(gel(P,2),T,p));
-  int t=find_size(n)+1;
-  int tab[t];
+  long t=find_size(n)+1;
+  long tab[t];
   find_list(tab,n);
-  int k=0;
+  long k=0;
   while(k<t-1)
   {	
     ret=FqX_Newton_iteration_inv(ret,P,tab[k],tab[k+1],T,p);
@@ -2730,7 +2730,7 @@ GEN FqX_primitive(GEN P,GEN T,GEN p)
   }
   GEN ret=cgetg(lg(P)+1,t_POL);
   gel(ret,2)=Fq_ui(0,T,p);
-  int k=1;
+  long k=1;
   while(k<=lg(P)-2)
   {
     gel(ret,2+k)=Fq_mul(Fq_inv(Fq_ui(k,T,p),T,p),gel(P,2+k-1),T,p);
@@ -2739,7 +2739,7 @@ GEN FqX_primitive(GEN P,GEN T,GEN p)
   return ret;
 	
 }
-GEN FqXn_log(GEN g,int n,GEN T,GEN p)
+GEN FqXn_log(GEN g,long n,GEN T,GEN p)
 {
   GEN res;
   res=FqXn_mul(FqX_deriv(g,T,p),FqXn_inv(g,n-1,T,p),n-1,T,p);
@@ -2747,7 +2747,7 @@ GEN FqXn_log(GEN g,int n,GEN T,GEN p)
   return res;
 }
 static
-GEN FqX_Newton_log_iteration(GEN g,GEN ginv,int t1,int t2,GEN T,GEN p)
+GEN FqX_Newton_log_iteration(GEN g,GEN ginv,long t1,long t2,GEN T,GEN p)
 {
   GEN res;	
   res=FqXn_mul(FqX_deriv(g,T,p),ginv,t2-1,T,p);
@@ -2755,22 +2755,22 @@ GEN FqX_Newton_log_iteration(GEN g,GEN ginv,int t1,int t2,GEN T,GEN p)
   return res;
 }
 static
-GEN FqX_Newton_iteration_exp(GEN g,GEN ginv,GEN f,int t1,int t2,GEN T,GEN p)
+GEN FqX_Newton_iteration_exp(GEN g,GEN ginv,GEN f,long t1,long t2,GEN T,GEN p)
 //assumes the inverse of g is known as ginv=1/g mod x^(t2-1) and g=exp(f) mod x^t1 
 //returns exp(f) mod x^t2
 {
   return FqX_add(g,RgX_shift(FqXn_mul(g,RgX_shift(FqX_sub(f,FqX_Newton_log_iteration(g,ginv,t1,t2,T,p),T,p),-t1),t2-t1,T,p),t1),T,p);
 }
-GEN FqXn_exp(GEN f,int n,GEN T,GEN p)
+GEN FqXn_exp(GEN f,long n,GEN T,GEN p)
 {
   GEN ret;
   GEN i;
   ret=mkpoln(1,Fq_ui(1,T,p));
   i=mkpoln(1,Fq_ui(1,T,p));
-  int t=find_size(n)+1;
-  int tab[t];
+  long t=find_size(n)+1;
+  long tab[t];
   find_list(tab,n);
-  int j=0;
+  long j=0;
   while(j<t-1)
   {	
   //we only have about (>=) 1/4 of the precision of the inverse at each step.(One could expect 1/2)
